multiple changes
Added alert which handles bugs when invalid input is entered, also
added tag values so alertView can respond to different alerts.
Added new keyboard bar, which made the done and cancel buttons that
were in the interface unnecessary. also made textField the first
responder, without allowing an exit of the keyboard, so the keyboard
will be on the screen at all times